backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
motor==3.6.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.9
pydantic==2.9.2
openpyxl==3.1.5
email-validator==2.2.0
email-validator==2.2.0
